Automatically reconnect to database mid-request
So far this is just infrastructure: no errors are considered retryable,
and no real queries are marked retryable.

Even in this state, we can drop connection verification from checkout,
and defer it until the first query is ready. But the more interesting
within-reach win will be when 'BEGIN' is retryable, and we can recognise
connectivity-related exceptions.